Avoid error when stylesheet doesn't have an href
I am getting an error when the href on a stylesheet is null. The current code is checking for an empty string but not null. This patch fixes this.

For an example use case I am getting null on the href via the stylesheet tag generated by the htm5shiv.js which adds support for HTML5 elements to older versions of IE.